feat: rename to core
